Замена half на float для всех шейдеров

+++ shaders/r1/common_defines.h
@@ -28,6 +28,6 @@
 #define	MAX_SHIFT		0.011f
 #define RANGE			100.f
 
-#define def_distort	half(0.05f)	// we get -0.5 .. 0.5 range, this is -512 .. 512 for 1024, so scale it
+#define def_distort	float(0.05f)	// we get -0.5 .. 0.5 range, this is -512 .. 512 for 1024, so scale it
 ////////////////////////////////////////////////////////////////////////////
 #endif	//	COMMON_DEFINES_H_INCLUDED
